+++ docker-compose.yml
@@ -11,11 +11,11 @@ services:
     ports:
       - ${POSTGRES_PORT}:5432
 
-  hoa-event:
+  hoa-webhook:
     build:
       context: . # Gradle requires the root context since this is a multi-module gradle build.
-      dockerfile: ./event/test.dockerfile
-    image: hoa-event:latest
+      dockerfile: ./webhook/test.dockerfile
+    image: hoa-webhook:latest
     environment:
       - POSTGRES_PASSWORD=${POSTGRES_PASSWORD}
       - POSTGRES_USER=${POSTGRES_USER}
@@ -24,4 +24,4 @@ services:
       - POSTGRES_PORT=${POSTGRES_PORT}
       - LUMA_API_KEY=${LUMA_API_KEY}
     ports:
-      - ${EVENT_PORT}:8080
+      - ${WEBHOOK_PORT}:8080
