--- event/run-jar.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-set -e
-set -u
-set -o pipefail
-
-java -jar /usr/local/bin/event-0.0.1-all.jar
