--- event/src/main/kotlin/org/climatechangemakers/hoa/event/LumaEvent.kt
@@ -1,13 +1,0 @@
-package org.climatechangemakers.hoa.event
-
-import kotlinx.datetime.Instant
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable data class LumaEvent(
-  @SerialName("api_id") val id: String,
-  val name: String,
-  @SerialName("secret_key") val secret: String?,
-  @SerialName("start_at") val startTime: Instant,
-  @SerialName("duration_minutes") val durationMinutes: Int?,
-)
