--- event/src/test/kotlin/org/climatechangemakers/hoa/event/FakeClock.kt
@@ -1,8 +1,0 @@
-package org.climatechangemakers.hoa.event
-
-import kotlinx.datetime.Clock
-import kotlinx.datetime.Instant
-
-class FakeClock(private val now: Instant) : Clock {
-  override fun now(): Instant = now
-}
