+++ webhook/src/test/kotlin/org/climatechangemakers/hoa/webhook/FakeClock.kt
@@ -1,0 +1,8 @@
+package org.climatechangemakers.hoa.webhook
+
+import kotlinx.datetime.Clock
+import kotlinx.datetime.Instant
+
+class FakeClock(private val now: Instant) : Clock {
+  override fun now(): Instant = now
+}
